+++ JkrLuaExe/JkrLuaExe.cpp
@@ -13,6 +13,7 @@ extern void CreateTextRendererBindings(sol::state& inState);
 extern void CreateBuildSystemBindings(sol::state& inS);
 extern void CreateRenderer3DBindings(sol::state& s);
 extern umap<s, v<s>> CommandLine(int ArgCount, char** ArgStrings); // TODO Complete This
+extern void CreateMultiThreadingBindings(sol::state& inState);
 
 void CreateMainBindings(sol::state& s) {
           s.open_libraries();
@@ -22,6 +23,7 @@ void CreateMainBindings(sol::state& s) {
           CreateRendererBindings(s);
           CreateTextRendererBindings(s);
           CreateMiscBindings(s);
+          CreateMultiThreadingBindings(s);
           CreateRenderer3DBindings(s);
 }
 } // namespace JkrEXE
