+++ JkrLuaExe/JkrLuaExe.hpp
@@ -1,10 +1,29 @@
 ﻿// JkrLuaExe.h : Header file for your target.
 
 #pragma once
+#include "Pipeline/VulkanPipelineContext.hpp"
+#include "Pipeline/VulkanShaderModule.hpp"
+#include "Renderers/Renderer_base.hpp"
+#include "Renderers/ThreeD/Shape3D.hpp"
+#include "Renderers/TwoD/Line.hpp"
+#include "Renderers/TwoD/Shape.hpp"
+#include "Window.hpp"
+#include "glm/fwd.hpp"
+#include "ksai_config.hpp"
+#include "ksai_thread.hpp"
+#include "sol/sol.hpp"
 #include <EventManager.hpp>
 #include <Instance.hpp>
+#include <Misc/RecycleBin.hpp>
+#include <Renderers/BestText_Alt.hpp>
+#include <Renderers/ThreeD/Simple3D.hpp>
+#include <Renderers/TwoD/Shape.hpp>
 #include <WindowMulT.hpp>
+#include <functional>
+#include <glm/gtc/matrix_transform.hpp>
+#include <memory>
 #include <sol/sol.hpp>
+#include <vector>
 
 sol::state& GetMainStateRef();
 
@@ -27,4 +46,8 @@ static auto my_exception_handler(lua_State* L, sol::optional<const std::exceptio
                     std::cout << std::endl;
           }
           return sol::stack::push(L, description);
+};
+
+struct DefaultCustomImagePainterPushConstant {
+          glm::vec4 x, y, z;
 };
