+++ JkrLuaExe/JkrMath.cpp
@@ -1,6 +1,4 @@
 #include "JkrLuaExe.hpp"
-#include "glm/fwd.hpp"
-#include <glm/gtc/matrix_transform.hpp>
 
 namespace JkrEXE {
